Update import paths to use @src/types instead of @ic-reactor/core/dist/types
