--- packages/core/src/classes/candid/parser.d.ts
@@ -1,48 +1,0 @@
-declare module '@ic-reactor/parser' {
-	/* tslint:disable */
-	/* eslint-disable */
-	/**
-	* @param {string} prog
-	* @returns {string}
-	*/
-	export function did_to_js(prog: string): string;
-	/**
-	* @param {string} prog
-	* @returns {string}
-	*/
-	export function did_to_ts(prog: string): string;
-
-	export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembly.Module;
-
-	export interface InitOutput {
-	  readonly memory: WebAssembly.Memory;
-	  readonly did_to_js: (a: number, b: number, c: number) => void;
-	  readonly did_to_ts: (a: number, b: number, c: number) => void;
-	  readonly __wbindgen_add_to_stack_pointer: (a: number) => number;
-	  readonly __wbindgen_malloc: (a: number, b: number) => number;
-	  readonly __wbindgen_realloc: (a: number, b: number, c: number, d: number) => number;
-	  readonly __wbindgen_free: (a: number, b: number, c: number) => void;
-	}
-
-	export type SyncInitInput = BufferSource | WebAssembly.Module;
-	/**
-	* Instantiates the given `module`, which can either be bytes or
-	* a precompiled `WebAssembly.Module`.
-	*
-	* @param {SyncInitInput} module
-	*
-	* @returns {InitOutput}
-	*/
-	export function initSync(module: SyncInitInput): InitOutput;
-
-	/**
-	* If `module_or_path` is {RequestInfo} or {URL}, makes a request and
-	* for everything else, calls `WebAssembly.instantiate` directly.
-	*
-	* @param {InitInput | Promise<InitInput>} module_or_path
-	*
-	* @returns {Promise<InitOutput>}
-	*/
-	export default function __wbg_init (module_or_path?: InitInput | Promise<InitInput>): Promise<InitOutput>;
-
-}
