+++ packages/parser/package.json
@@ -27,17 +27,13 @@
   "homepage": "https://github.com/b3hr4d/ic-reactor/tree/main/packages/parser#readme",
   "scripts": {
     "test": "NODE_OPTIONS=\"$NODE_OPTIONS --experimental-vm-modules\" npx jest",
-    "build": "wasm-pack build --release --target web --out-dir dist --out-name index && node ./dts-generator.js",
+    "build": "wasm-pack build --release --target web --out-dir dist --out-name index",
     "build:node": "wasm-pack build --release --target nodejs --out-dir dist --out-name index",
     "build:builder": "wasm-pack build --release --out-dir dist --out-name index",
     "start": "tsc watch",
-    "clean": "npx rimraf dist",
-    "generate-types": "node ./dts-generator.js"
+    "clean": "npx rimraf dist"
   },
   "engines": {
     "node": ">=10"
-  },
-  "devDependencies": {
-    "dts-generator": "^3.0.0"
   }
 }
