+++ packages/visitor/src/details/types.ts
@@ -1,4 +1,4 @@
-import type { BaseActor, FunctionName } from "@ic-reactor/core/dist/types"
+import type { BaseActor, FunctionName } from "@src/types"
 import type { MethodReturnDetails } from "./returns/types"
 import type { MethodArgDetails } from "./args/types"
