+++ packages/visitor/src/fields/types.ts
@@ -1,4 +1,4 @@
-import type { BaseActor, FunctionName } from "@ic-reactor/core/dist/types"
+import type { BaseActor, FunctionName } from "@src/types"
 import type { MethodReturns } from "./returns/types"
 import type { MethodArgs } from "./args/types"
