+++ packages/visitor/src/index.ts
@@ -5,5 +5,5 @@ export * from "./transform"
 export * from "./random/args"
 export * from "./random/response"
 
-export * as helper from "./helper"
+export * as helper from "./helpers"
 export * as types from "./types"
