--- packages/visitor/src/layouts/constants.ts
@@ -1,119 +1,0 @@
-export const WALLET_TEST = [
-  "wallet",
-  "balance",
-  "transfer",
-  "send",
-  "topup",
-  "receive",
-  "deposit",
-  "withdraw",
-  "pay",
-  "claim",
-  "refund",
-  "reward",
-  "tip",
-  "buy",
-  "sell",
-  "purchase",
-  "order",
-  "checkout",
-  "cart",
-  "invoice",
-  "payment",
-  "credit",
-  "debit",
-  "transaction",
-  "history",
-  "statement",
-  "account",
-  "address",
-  "fund",
-  "staking",
-  "bond",
-  "unbond",
-  "delegate",
-  "undelegate",
-  "approve",
-  "allowance",
-]
-
-export const SETTING_TEST = [
-  "setting",
-  "version",
-  "set",
-  "update",
-  "change",
-  "modify",
-  "edit",
-  "remove",
-  "delete",
-  "add",
-  "create",
-  "clear",
-  "reset",
-  "revoke",
-  "renew",
-  "replace",
-  "upgrade",
-  "load",
-  "upload",
-  "downgrade",
-  "install",
-  "role",
-  "uninstall",
-  "enable",
-  "disable",
-  "activate",
-  "deactivate",
-  "suspend",
-  "resume",
-  "pause",
-  "start",
-  "stop",
-  "restart",
-  "refresh",
-  "reload",
-  "reboot",
-  "shutdown",
-  "terminate",
-  "kill",
-  "abort",
-  "cancel",
-]
-
-export const STATUS_TEST = [
-  "status",
-  "state",
-  "condition",
-  "phase",
-  "step",
-  "stage",
-  "level",
-  "wasm",
-  "detail",
-  "info",
-  "message",
-  "note",
-  "log",
-  "bug",
-  "error",
-  "capacity",
-  "partition",
-  "timer",
-]
-
-export const GOVERNANCE_TEST = ["governance", "vote", "poll", "proposal"]
-
-export const DEFAULT_LAYOUTS = [
-  { name: "xl", size: 6 },
-  { name: "md", size: 4 },
-  { name: "xs", size: 2 },
-] as const
-
-export const DEFAULT_CATEGORIES = [
-  "home",
-  "wallet",
-  "governance",
-  "setting",
-  "status",
-] as const
