+++ packages/visitor/src/transform/helpers.ts
@@ -1,4 +1,4 @@
-import { MethodResult } from "./types"
+import type { MethodResult } from "./types"
 
 export const isValueInTable = (value: MethodResult): boolean => {
   if (value.type === "optional" && value.value) {
